compute progress
Former-commit-id: 713a5a48d75d760ee55d489611db9b9697e559cf

+++ src/main/java/cz/cvut/fel/aic/amodsim/ridesharing/vga/calculations/GroupData.java
@@ -17,7 +17,7 @@ public class GroupData {
     private final Set<PlanComputationRequest> requests;
 
     private final Set<PlanComputationRequest> onboardRequestLock;
-    private float feasible;
+    private double feasible;
     public GroupData(Set<PlanComputationRequest> requests) {
         this(requests, null);
     }
@@ -61,11 +61,11 @@ public class GroupData {
         return onboardRequestLock;
     }
 
-    public float getFeasible() {
+    public double getFeasible() {
         return feasible;
     }
 
-    public void setFeasible(float feasible) {
+    public void setFeasible(double feasible) {
         this.feasible = feasible;
     }
